nlp-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
konlpy==0.6.0
wordcloud==1.9.3
matplotlib==3.8.4
numpy==1.26.4
pillow==10.3.0